move general frontier related functions to own files and introduce `DoFacilityTrainerBattle`
