+++ include/battle_dome.h
@@ -1,6 +1,8 @@
 #ifndef GUARD_BATTLE_DOME_H
 #define GUARD_BATTLE_DOME_H
 
+#include "constants/battle_dome.h"
+
 int GetDomeTrainerSelectedMons(u16 tournamentTrainerId);
 int TrainerIdToDomeTournamentId(u16 trainerId);
