+++ include/battle_tower.h
@@ -2,6 +2,7 @@
 #define GUARD_BATTLE_TOWER_H
 
 #include "data.h"
+#include "battle_frontier.h"
 
 struct RSBattleTowerRecord
 {
@@ -15,17 +16,6 @@ struct RSBattleTowerRecord
     /*0xA0*/ u32 checksum;
 };
 
-struct BattleFrontierTrainer
-{
-    u8 facilityClass;
-    u8 filler1[3];
-    u8 trainerName[PLAYER_NAME_LENGTH + 1];
-    u16 speechBefore[EASY_CHAT_BATTLE_WORDS_COUNT];
-    u16 speechWin[EASY_CHAT_BATTLE_WORDS_COUNT];
-    u16 speechLose[EASY_CHAT_BATTLE_WORDS_COUNT];
-    const u16 *monSet;
-};
-
 extern const u8 gTowerMaleFacilityClasses[30];
 extern const u16 gTowerMaleTrainerGfxIds[30];
 extern const u8 gTowerFemaleFacilityClasses[20];
@@ -35,26 +25,9 @@ extern const struct BattleFrontierTrainer gBattleFrontierTrainers[];
 extern const struct TrainerMon gSlateportBattleTentMons[];
 extern const struct BattleFrontierTrainer gSlateportBattleTentTrainers[];
 
-// Temporary storage for monIds of the opponent team
-// during team generation in battle factory and similar facilities.
-extern u16 gFrontierTempParty[];
-
-extern const struct BattleFrontierTrainer *gFacilityTrainers;
-extern const struct TrainerMon *gFacilityTrainerMons;
-
 void CallBattleTowerFunc(void);
-u16 GetRandomScaledFrontierTrainerId(u8 challengeNum, u8 battleNum);
-void SetBattleFacilityTrainerGfxId(u16 trainerId, u8 tempVarId);
 void SetEReaderTrainerGfxId(void);
-u16 GetBattleFacilityTrainerGfxId(u16 trainerId);
 void PutNewBattleTowerRecord(struct EmeraldBattleTowerRecord *newRecordEm);
-u8 GetFrontierTrainerFrontSpriteId(u16 trainerId);
-u8 GetFrontierOpponentClass(u16 trainerId);
-void GetFrontierTrainerName(u8 *dst, u16 trainerId);
-void FillFrontierTrainerParty(u8 monsCount);
-void FillFrontierTrainersParties(u8 monsCount);
-u16 GetRandomFrontierMonFromSet(u16 trainerId);
-void FrontierSpeechToString(const u16 *words);
 void DoSpecialTrainerBattle(void);
 void CalcEmeraldBattleTowerChecksum(struct EmeraldBattleTowerRecord *record);
 void CalcRubyBattleTowerChecksum(struct RSBattleTowerRecord *record);
@@ -70,12 +43,10 @@ bool32 RubyBattleTowerRecordToEmerald(struct RSBattleTowerRecord *src, struct Em
 bool32 EmeraldBattleTowerRecordToRuby(struct EmeraldBattleTowerRecord *src, struct RSBattleTowerRecord *dst);
 void CalcApprenticeChecksum(struct Apprentice *apprentice);
 void GetBattleTowerTrainerLanguage(u8 *dst, u16 trainerId);
-u8 SetFacilityPtrsGetLevel(void);
-u8 GetFrontierEnemyMonLevel(u8 lvlMode);
-s32 GetHighestLevelInPlayerParty(void);
-u16 FacilityClassToGraphicsId(u8 facilityClass);
+u8 SetTentPtrsGetLevel(void);
 bool32 ValidateBattleTowerRecord(u8 recordId); // unused
 void TrySetLinkBattleTowerEnemyPartyLevel(void);
-void CreateFacilityMon(const struct TrainerMon *fmon, u16 level, u8 fixedIV, u32 otID, u32 flags, struct Pokemon *dst);
+void FillTentTrainerParty(u8 monsCount);
+
 
 #endif //GUARD_BATTLE_TOWER_H
