+++ src/battle_controller_recorded_opponent.c
@@ -10,6 +10,7 @@
 #include "battle_tv.h"
 #include "bg.h"
 #include "data.h"
+#include "frontier_util.h"
 #include "item_menu.h"
 #include "item_use.h"
 #include "link.h"
